pyproject.toml
[tool.poetry]
name = "sigbox-lever-video"
version = "0.1.0"
description = ""
authors = ["Lowdham Railway Heritage"]
package-mode = false

[tool.poetry.dependencies]
python = ">=3.13,<3.14"
# In some projects, this fails with a virus false positive
#pyinstaller = "^6.7"
fastapi = "^0.115"
uvicorn = "^0.34"
python-vlc = "^3.0"

[tool.poetry.group.dev.dependencies]
black = "^25"
mypy = "^1.14"
pre-commit = "^4"
psutil = "^6.1"
isort = "^6"
bump2version = "^1.0"
setuptools = "^70"

[tool.black]
line_length = 98

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
line_length = 98

[tool.coverage.report]
include = ['./source/*']
exclude_lines = ["if __name__ == .__main__.:"]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
